feat: Implement basic car racing game flow
- Prompt the user to input car names seperated by commas
- Create instances of Car based on the input
- Move each car once
- Display car names and their progress
- Seperate UI print logic to ui/ouput.js
