+++ docs/REQUIREMENTS.md
@@ -2,16 +2,17 @@
 
 #### 자동차 이름 입력받기
 
-- [] 자동차 이름은 쉼표(,)로 구분
-  - [] 각 이름의 앞뒤 공백 제거
+- [x] 자동차 이름은 쉼표(,)로 구분
+  - [x] 각 이름의 앞뒤 공백 제거
+  - [x] 입력받은 이름을 바탕으로 자동차 객체 생성
 - [] 각 자동차 이름은 1자 이상 5자 이하
 - [] 중복된 자동차 이름을 입력할 수 없음
 
 #### 자동차 경주 진행
 
 - [] 자동차 경주 횟수는 5회로 고정
-- [] 각 라운드마다 자동차가 한칸 씩 전진 (`-` 추가)
-- [] 각 라운드 종료 후 자동차 궤적 출력
+- [x] 각 라운드마다 자동차가 한칸 씩 전진 (`-` 추가)
+- [x] 각 라운드 종료 후 자동차 궤적 출력
 - [] 모든 라운드 완료 시 "경주를 완료했습니다" 출력 후 게임 종료
 
 #### 게임 종료
